sound for windowbars, see description
FMODUnity.RuntimeManager.PlayOneShot ("event:/fx/windowBars");
This sound is for one bar segment, so it should be synchronized with each of the four segments on the left and the three on the right. It should be correct length to be launched when a segment starts to move. I will polish the sound after I can listen to it in game!